% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmin.R
\name{backtrack}
\alias{backtrack}
\title{Backtrack to find a new candidate step size}
\usage{
backtrack(alpha, fvals, gval, minstep = 0.1, ...)
}
\arguments{
\item{alpha}{previous 2 step sizes}

\item{fvals}{previous 2 functional values for alpha step sizes
(or 1 if first substep)}

\item{minstep}{minimum relative step size change}

\item{...}{other arguments to f}
}
\value{
proposed step size
}
\description{
See p58 of Nocedal and Wright (2006).
}
