% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmin.R
\name{check_fmin}
\alias{check_fmin}
\title{Check optimization from data stored during iterations in fmin}
\usage{
check_fmin(opt)
}
\arguments{
\item{opt}{output of fmin when fmin is run with save = TRUE argument}
}
\value{
four plots: norm of difference between parameters and final estimates
                    value of objective function over iterations
                    norm of gradient over iterations
                    norm of relative gradient over iterations
}
\description{
Check optimization from data stored during iterations in fmin
}
