% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmin.R
\name{check_positive_definite}
\alias{check_positive_definite}
\title{Check if matrix is positive definite}
\usage{
check_positive_definite(X, tol = 1e-10)
}
\arguments{
\item{X}{matrix}

\item{tol}{all eigenvalues less than tol in absolute value are rounded to zero}
}
\value{
TRUE if positive definite, FALSE otherwise
}
\description{
Check if matrix is positive definite
}
