% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmin.R
\name{get_halfstep}
\alias{get_halfstep}
\title{Do a partial Newton step with line search selected step size}
\usage{
get_halfstep(alpha, theta, f, der, fval, fvals, newton_step, halfsteps,
  ...)
}
\arguments{
\item{alpha}{previous 2 step sizes  (or 1 if first halfstep)}

\item{theta}{current iterateo}

\item{f}{objective function}

\item{der}{directional derviative of f in direction of gradient g}

\item{fval}{f(theta) current function value}

\item{fvals}{previous 2 functional values (or 1 if first halfstep)}

\item{newton_step}{full Newton update step}

\item{halfsteps}{number of halfsteps taken so far}

\item{...}{other arguments to f}
}
\value{
vector with two most recent step sizes (entries 1, 2) and two most recent function evaluations (3, 4)
}
\description{
Do a partial Newton step with line search selected step size
}
