% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmin.R
\name{get_newton_step}
\alias{get_newton_step}
\title{Do a Newton update with a Cholesky factor for the Hessian}
\usage{
get_newton_step(R, g)
}
\arguments{
\item{R}{Cholesky factor (upper triangular) for the Hessian}

\item{g}{gradient vector}
}
\value{
Updated Cholesky
}
\description{
Do a Newton update with a Cholesky factor for the Hessian
}
