% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmin.R
\name{linesearch}
\alias{linesearch}
\title{Perform line search in direction of newton step}
\usage{
linesearch(
  theta,
  newton_step,
  fval,
  der,
  f,
  gfn,
  maxsubsteps,
  stepmax,
  funit,
  units,
  ...
)
}
\arguments{
\item{theta}{current function values}

\item{newton_step}{full newton step}

\item{fval}{current function value}

\item{der}{directional derivative of f in direction of newton_step}

\item{f}{objective function}

\item{gfn}{gradient function}

\item{maxsubsteps}{maximum substeps to perform}

\item{stepmax}{maximum step size allowed}

\item{...}{other arguments to f}
}
\description{
Perform line search in direction of newton step
}
