% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmin.R
\name{make_positive_definite}
\alias{make_positive_definite}
\title{Add epsilon to diagonal of matrix until it is positive definite}
\usage{
make_positive_definite(X, epsilon = 1e-07, maxit = 100, tol = 1e-10)
}
\arguments{
\item{X}{a square matrix}

\item{epsilon}{amount to add on each iteration}

\item{maxit}{maximum number of iterations to try}

\item{tol}{tolerance for check_positive_definite}
}
\value{
altered matrix
}
\description{
Add epsilon to diagonal of matrix until it is positive definite
}
