% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmin.R
\name{zoom}
\alias{zoom}
\title{Zoom in on step size given interval where Wolfe conditions can be satisfied}
\usage{
zoom(
  alpha,
  theta,
  f,
  gfn,
  der,
  fval,
  fvals,
  gval,
  newton_step,
  maxsubsteps,
  ...
)
}
\arguments{
\item{alpha}{vector of 2 step sizes, boundary of interval}

\item{theta}{current parameter values}

\item{f}{objective function}

\item{gfn}{gradient function}

\item{der}{direction derivative of f in Newton direction}

\item{fval}{current function value}

\item{fvals}{function values at boundary step sizes}

\item{newton_step}{full Newton step direction}

\item{maxsubsteps}{maximum number of substeps allowed}

\item{...}{other arguments to f}
}
\description{
Zoom in on step size given interval where Wolfe conditions can be satisfied
}
