% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmin.R
\name{bfgs_update}
\alias{bfgs_update}
\title{Update Cholesky factor of the Hessian by BFGS method}
\usage{
bfgs_update(R, gdif, delta)
}
\arguments{
\item{R}{Cholesky factor}

\item{gdif}{change in gradients}

\item{delta}{step taken}
}
\value{
Updated Cholesky factor R
}
\description{
Update Cholesky factor of the Hessian by BFGS method
}
