% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmin.R
\name{fmin}
\alias{fmin}
\title{Minimize a function f using a Quasi-Newton method}
\usage{
fmin(
  obj,
  theta,
  gobj = NULL,
  funit = 1,
  units = NULL,
  maxit = 1000,
  tol = 1e-10,
  stepmax = 1,
  maxsubsteps = 10,
  save = FALSE,
  verbose = FALSE,
  digits = 4,
  ...
)
}
\arguments{
\item{obj}{function to be minimised which takes parameter vector as
first argument}

\item{gobj}{(optional) gradient function of f, uses finite differencing
otherwise}

\item{funit}{scaling for function values, try to pick this so that f 
values vary between -1 and 1}

\item{units}{a vector of same length as start giving coordinate units, 
try to pick units such that values of parameter in that
coordinate direction vary between -1 and 1}

\item{maxit}{maximum number of iterations to try}

\item{tol}{tolerance for stopping criteria}

\item{stepmax}{maximum relative step length, default of 1 means no steps
longer than full Newton step}

\item{maxsubsteps}{maximum number of substeps in line search}

\item{save}{if TRUE then parameter values, function values, and gradients
are saved for every iteration}

\item{verbose}{if TRUE, function value and then parameter value are printed
for each iteration}

\item{digits}{number of digits to print when verbose = TRUE}

\item{...}{other named arguments to f and gobj (if given)}

\item{start}{vector of starting values}
}
\value{
a list with elements:
\itemize{
  \item estimate - a vector of the optimal estimates
  \item value -  value of the objective function at the optimum
  \item g - gradient vector at the optimum
  \item H - hessian at the optimum computed by finite difference
  \item conv - is TRUE if convergence was satisfied, otherwise may not have
                converged on optimum
  \item niter: number of iterations taken
}
}
\description{
Performs a quasi-Newton line search optimization with BFGS
             updates of the inverse hessian. See Nocedal and Wright
             (2006), Chapter 3 for details.
}
