% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmin.R
\name{get_newton_step}
\alias{get_newton_step}
\title{Do a Newton update with inverse Hessian}
\usage{
get_newton_step(H, g)
}
\arguments{
\item{H}{Inverse Hessian matrix}

\item{g}{gradient vector}
}
\value{
Newton step vector
}
\description{
Do a Newton update with inverse Hessian
}
